+++ packages/starlight-theme-nova/src/shiki-config.ts
@@ -10,7 +10,6 @@ import type { ShikiTransformer } from '@shikijs/types'
 import type { ShikiConfig } from 'astro'
 import { createRenderer } from 'shiki-twoslash-renderer'
 
-import { transformerMetaTitle } from './shiki-tranformer-meta-title'
 import { transformerContainer } from './shiki-transformer-container'
 import { transformerCopyButton } from './shiki-transformer-copy-button'
 
@@ -23,7 +22,6 @@ export function createShikiConfig(options: { twoslash: boolean }): ShikiConfig {
     transformerNotationWordHighlight(),
 
     transformerCopyButton(),
-    transformerMetaTitle(),
     transformerContainer(),
     options.twoslash
       ? transformerTwoslash({
