+++ packages/starlight-theme-nova/src/shiki-meta-title.ts
@@ -1,0 +1,6 @@
+export function parseTitleString(meta: string): string | null {
+  if (!meta) return null
+  const match = meta.match(/title="([^"]+)"/)
+  if (!match) return null
+  return match[1]
+}
